savestate, good working version
This version does not suffer the "dancing" problem as often, and will report that it has completed its goals.
Also increased costmap potential. The nav2 docs reccommend having a subtle cost curve throughout the map as this improves the performance of the planner.
Decreasing the accel limits in the vel smoother config seem to reduce slippage inside the lab.